Fix excessive interpolation of missing data in Song.get_plays()
Song.get_plays() linearly interpolates the raw streaming data to return
data points at the top of the hour (i.e. 14:00 as opposed to 14:17). The
interpolation as previously implemented didn't differentiate between
missing data in between two adjacent measurements and missing data
caused by a failure of the fetching process. The new implementation
ensures preservation of missing data in the original data set, at the
cost of the method now returning a Series of floats rather than a Series
of integers. This cannot be currently fixed, as Pandas doesn't support
missing values in arrays of integers. In any case this shouldn't be more
than a minor nuisance, as the order of magnitudes involved is not big
enough to cause loss of precision due to rounding.